+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## x
 - Edit version of junit: from 3.8.1 to 4.12;
+- use assertTrue/assertFalse instead of assertEquals on boolean values;
 
 
 ## 0.1.0